refactor: Renamed `CheckboxCollapse.tsx` to `ToggleCollapse.tsx`
